feat(go-sdk): :sparkles: 同步 SDK v1.18.2 → v1.18.4
- 清理 config.go 中废弃的 permission 类型和 config mode 枚举
- 统一 Go 命名规范：Api→API，type alias 对齐 OpenAPI schema 名称
- 新增 SubtaskPartModelParam、PermissionV2SourceParam 等请求参数类型
- provider.go 中 Interleaved 字段改为 Union 类型，float64→int64 修正
- v2reference.go 将 ReferenceSource 改为具体包装类型
- 修复 question.go 参数验证顺序、v2fs.go 添加路径空值校验
- 完善 formatter.go/lsp.go 等文件注释文档

+++ experimental_session.go
@@ -96,7 +96,7 @@ type GlobalSession struct {
 	Title     string            `json:"title,required"`
 	Version   string            `json:"version,required"`
 	Time      GlobalSessionTime `json:"time,required"`
-	// This field can have the runtime type of [ProjectSummary].
+	// This field can have the runtime type of [ProjectSummary] or nil.
 	Project     interface{}          `json:"project,required"`
 	WorkspaceID string               `json:"workspaceID,omitempty"`
 	Path        string               `json:"path,omitempty"`
@@ -153,7 +153,7 @@ type GlobalSessionTime struct {
 	Created    int64                 `json:"created,required"`
 	Updated    int64                 `json:"updated,required"`
 	Compacting int64                 `json:"compacting,omitempty"`
-	Archived   int64                 `json:"archived,omitempty"`
+	Archived   float64               `json:"archived,omitempty"`
 	JSON       globalSessionTimeJSON `json:"-"`
 }
