+++ provider_test.go
@@ -74,7 +74,7 @@ func TestProviderOauthAuthorize(t *testing.T) {
 		option.WithBaseURL(baseURL),
 	)
 	_, err := client.Provider.OauthAuthorize(context.TODO(), "providerID", opencode.ProviderOauthAuthorizeParams{
-		Method: opencode.F(0.000000),
+		Method: opencode.F(int64(0)),
 	})
 	if err != nil {
 		var apierr *opencode.Error
@@ -98,7 +98,7 @@ func TestProviderOauthCallback(t *testing.T) {
 		option.WithBaseURL(baseURL),
 	)
 	_, err := client.Provider.OauthCallback(context.TODO(), "providerID", opencode.ProviderOauthCallbackParams{
-		Method: opencode.F(0.000000),
+		Method: opencode.F(int64(0)),
 		Code:   opencode.F("code"),
 	})
 	if err != nil {
